@BohuTANG 2021-03-01
Test this patch on my ubuntu:
before:
```
mysql> select avg(number) from system.numbers_mt(50000000000);
+--------------------+
| Avg(number)        |
+--------------------+
| 281362940.72920084 |
+--------------------+
1 row in set (1.55 sec)

mysql> select sum(number) from system.numbers_mt(50000000000);
+----------------------+
| Sum(number)          |
+----------------------+
| 14068147036460041728 |
+----------------------+
1 row in set (1.46 sec)
```

after:

```
mysql> select sum(number) from system.numbers_mt(50000000000);
+----------------------+
| Sum(number)          |
+----------------------+
| 14068147036460041728 |
+----------------------+
1 row in set (1.48 sec)

mysql> select avg(number) from system.numbers_mt(50000000000);
+--------------------+
| Avg(number)        |
+--------------------+
| 281362940.72920084 |
+--------------------+
1 row in set (1.53 sec)
```
No performance improvement.